requirements.txt
# Core dependencies
numpy>=1.24.0
scipy>=1.10.0

# Deep learning (for comparison and GPU support)
torch>=2.0.0
torchvision>=0.15.0

# Data processing
pandas>=2.0.0
scikit-learn>=1.3.0
pillow>=10.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0

# API and deployment
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
requests>=2.31.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code quality
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0
